useNotificationPoller hook for retrieving notifications, adjustments tied to the move of notification drawer component to a prop of Page, removal of unnecessary components
