+++ frontend/packages/patternfly/src/components/notification-drawer/index.ts
@@ -1,4 +1,2 @@
-export { default as NotificationDrawer } from './notification-drawer';
-export { default as NotificationDrawerHeading } from './notification-drawer-heading';
 export { default as NotificationEntry, NotificationTypes } from './notification-entry';
 export { default as NotificationCategory } from './notification-category';
